fixed imports after moving publicAPIV2 class
