+++ tests/test_base.py
@@ -1,9 +1,7 @@
 import mock
 
-from sxapi.base import (
-    IntegrationAPIV2,
-    PublicAPIV2,
-)
+from sxapi.base import IntegrationAPIV2
+from sxapi.publicV2 import PublicAPIV2
 
 
 @mock.patch("requests.Session.delete")
